Updated the data type of the funds field in the Account table along with updated UTs for the same

+++ Auction_backend/src/main/java/com/online/auction/controller/AccountController.java
@@ -10,7 +10,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import static com.online.auction.constant.AuctionConstants.ACCOUNT;
@@ -24,10 +26,10 @@ public class AccountController {
     private AccountService accountService;
 
     @GetMapping("/balance")
-    public ResponseEntity<SuccessResponse<Float>> getAccountBalance(
+    public ResponseEntity<SuccessResponse<Double>> getAccountBalance(
             @AuthenticationPrincipal User user) throws ServiceException {
-        float balance = accountService.getAccountBalance(user.getUserId());
-        SuccessResponse<Float> response = new SuccessResponse<>(200, HttpStatus.OK, balance);
+        double balance = accountService.getAccountBalance(user.getUserId());
+        SuccessResponse<Double> response = new SuccessResponse<>(200, HttpStatus.OK, balance);
         return ResponseEntity.ok(response);
 
     }
